fixed missing parameter days

+++ functions.php
@@ -37,8 +37,14 @@
 		$GLOBALS["use_db"] = 0;
 	}
 
-	function can_edit($id) {
+	function can_edit($model_user_id) {
+		$user_id = get_user_id_from_session_id($_COOKIE["session_id"]);
+		_assert($user_id > 0, "id is too low");
 		if(get_single_value_from_query("select role_id from login where id = ".$id) == 1){
+			return true;
+		}
+		//$model_user_id = 101;
+		if($user_id == $model_user_id) {
 			return true;
 		}
 		return false;
@@ -59,10 +65,12 @@
 	}
 
 	function delete_mongo ($collection, $id) {
-		$bulk = new \MongoDB\Driver\BulkWrite();
-		$bulk->delete(array('_id' => new MongoDB\BSON\ObjectId($id)));
-		$result = $GLOBALS["manager"]->executeBulkWrite($collection, $bulk);
-		return $result;
+		if(can_delete($id)) {
+			$bulk = new \MongoDB\Driver\BulkWrite();
+			$bulk->delete(array('_id' => new MongoDB\BSON\ObjectId($id)));
+			$result = $GLOBALS["manager"]->executeBulkWrite($collection, $bulk);
+			return $result;
+		}
 	}
 
 	function delete_mongo_models ($id) {
